Stop changing direction of bezier curves; preserve horizontal lines.

+++ src/glyph.hpp
@@ -37,7 +37,7 @@ public:
 
     const GlyphInfo& info() const { return m_info; }
 
-    size_t isInside(ivec2 pos, size_t& beginAt) const noexcept;
+    size_t isInside(vec2 pos, size_t& beginAt) const noexcept;
 private:
 
     void extractOutlines(const std::vector<size_t>& contourEnd,
